requirements.txt
flask
groq
python-dotenv
pypdf
sentence-transformers
numpy
